Move sidenav to its own component
