+++ frontend_react/src/components/Sidenav/Sidenav.module.css
@@ -1,0 +1,19 @@
+.sideNav {
+  height: 100%;
+  width: 235px;
+  position: fixed;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  background-color: #52247F;
+}
+
+.logoContainer {
+  text-align: center;
+  width: 100%;
+  background-color: white;
+}
+
+.logo {
+  height: 100px;
+}
